+++ packages/header/tests/react/index.js
@@ -215,8 +215,8 @@ ReactDOM.render(
 					</div>
 				</AUheader>
 				<br />
-				<Route path="/tests/react/one" render={ () => ( <p>Route one</p> )} />
-				<Route path="/tests/react/two" render={ () => ( <p>Route two</p> )} />
+				<Route path="/tests/one" render={ () => ( <p>Route one</p> )} />
+				<Route path="/tests/two" render={ () => ( <p>Route two</p> )} />
 			</Fragment>
 		</BrowserRouter>