+++ packages/link-list/tests/react/index.js
@@ -100,8 +100,8 @@ ReactDOM.render(
 							text: 'Change to route two',
 						},
 					]} />
-					<Route path="/tests/react/one" render={ () => ( <p>Route one</p> )} />
-					<Route path="/tests/react/two" render={ () => ( <p>Route two</p> )} />
+					<Route path="/tests/one" render={ () => ( <p>Route one</p> )} />
+					<Route path="/tests/two" render={ () => ( <p>Route two</p> )} />
 				</Fragment>
 			</BrowserRouter>
 		</div>
@@ -198,8 +198,8 @@ ReactDOM.render(
 							text: 'Change to route two',
 						},
 					]} />
-					<Route path="/tests/react/one" render={ () => ( <p>Route one</p> )} />
-					<Route path="/tests/react/two" render={ () => ( <p>Route two</p> )} />
+					<Route path="/tests/one" render={ () => ( <p>Route one</p> )} />
+					<Route path="/tests/two" render={ () => ( <p>Route two</p> )} />
 				</Fragment>
 			</BrowserRouter>
 		</div>