+++ packages/inpage-nav/tests/react/index.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { BrowserRouter, Route, Link } from 'react-router-dom';
 
 import { AUinpageNavLinks, AUinpageNavSection } from './inpage-nav.js';